+++ plugins/postcss-custom-properties/index.js
@@ -115,7 +115,7 @@ function resolveValue(value, variables, source) {
   matches.body.replace(RE_VAR, function(_, name, fallback) {
     var replacement = variables[name]
     if (!replacement && !fallback) {
-      throw new Error(gnuMessage("variable '" + name + "' is undefined & don't have any fallback", source))
+      console.warn(gnuMessage("variable '" + name + "' is undefined & don't have any fallback", source))
     }
 
     // prepend with fallbacks